requirements.txt
gritlm==1.0.2
networkx==3.4.2
pydantic==2.10.4
python_igraph==0.11.8
tenacity==8.5.0
tiktoken==0.7.0
nest_asyncio
tqdm
einops
faiss-cpu
SQLAlchemy
sentence-transformers
gradio
arxiv 
ollama
pandas 
RapidFuzz
langchain-core
langchain-ollama
sqlglot
